+++ src/routes/issuer.js
@@ -2,7 +2,6 @@ import { Router } from "express";
 import { v4 as uuid } from "uuid";
 
 import issuerSchema from "../../ocf/schema/objects/Issuer.schema.json" assert { type: "json" };
-import issuerAuthorizedSharesAdjustment from "../../ocf/schema/objects/transactions/adjustment/IssuerAuthorizedSharesAdjustment.schema.json" assert { type: "json" };
 import deployCapTable from "../chain-operations/deployCapTable.js";
 import { createIssuer } from "../db/operations/create.js";
 import { countIssuers, readIssuerById } from "../db/operations/read.js";
@@ -11,7 +10,6 @@ import validateInputAgainstOCF from "../utils/validateInputAgainstSchema.js";
 
 import { contractCache } from "../utils/caches.js";
 import startOnchainListeners from "../chain-operations/transactionListener.js";
-import { convertAndAdjustIssuerAuthorizedSharesOnChain } from "../controllers/issuerController.js"
 
 const issuer = Router();
 
@@ -81,33 +79,6 @@ issuer.post("/create", async (req, res) => {
         console.log("Issuer created offchain:", issuer);
 
         res.status(200).send({ issuer });
-    } catch (error) {
-        console.error(`error: ${error}`);
-        res.status(500).send(`${error}`);
-    }
-});
-
-issuer.post("/adjust", async (req, res) => {
-    const { contract } = req;
-
-    try {
-        // OCF doesn't allow extra fields in their validation
-        const issuerAutnorizedSharesAdj = {
-            id: uuid(),
-            date: new Date().toISOString().slice(0, 10),
-            object_type: 'ISSUER', // "TX_ISSUER_AUTHORIZED_SHARES_ADJUSTMENT",
-            ...req.body,
-        };
-
-
-        await validateInputAgainstOCF(issuerAutnorizedSharesAdj,
-            issuerAuthorizedSharesAdjustment);
-
-        await convertAndAdjustIssuerAuthorizedSharesOnChain(contract,
-            ...issuerAutnorizedSharesAdj,
-        );
-
-        res.status(200).send({ issuerAutnorizedSharesAdj });
     } catch (error) {
         console.error(`error: ${error}`);
         res.status(500).send(`${error}`);
